# [Performance] Avoid double read file on parsing + create File object on creating StmtsAndTokens object