--- a/packages/FileSystemRector/Parser/FileInfoParser.php
+++ b/packages/FileSystemRector/Parser/FileInfoParser.php
@@ -29,9 +29,10 @@ final class FileInfoParser
      */
     public function parseFileInfoToNodesAndDecorate(string $filePath): array
     {
-        $stmts = $this->rectorParser->parseFile($filePath);
+        $fileContent = FileSystem::read($filePath);
+        $stmts = $this->rectorParser->parseString($fileContent);
 
-        $file = new File($filePath, FileSystem::read($filePath));
+        $file = new File($filePath, $fileContent);
         $stmts = $this->nodeScopeAndMetadataDecorator->decorateNodesFromFile($file, $stmts);
 
         $file->hydrateStmtsAndTokens($stmts, $stmts, []);