--- a/packages/Testing/TestingParser/TestingParser.php
+++ b/packages/Testing/TestingParser/TestingParser.php
@@ -32,8 +32,9 @@ final class TestingParser
         // needed for PHPStan reflection, as it caches the last processed file
         $this->dynamicSourceLocatorProvider->setFilePath($filePath);
 
-        $file = new File($filePath, FileSystem::read($filePath));
-        $stmts = $this->rectorParser->parseFile($filePath);
+        $fileContent = FileSystem::read($filePath);
+        $file = new File($filePath, $fileContent);
+        $stmts = $this->rectorParser->parseString($fileContent);
 
         $stmts = $this->nodeScopeAndMetadataDecorator->decorateNodesFromFile($file, $stmts);
 
@@ -53,8 +54,9 @@ final class TestingParser
 
         SimpleParameterProvider::setParameter(Option::SOURCE, [$filePath]);
 
-        $stmts = $this->rectorParser->parseFile($filePath);
-        $file = new File($filePath, FileSystem::read($filePath));
+        $fileContent = FileSystem::read($filePath);
+        $stmts = $this->rectorParser->parseString($fileContent);
+        $file = new File($filePath, $fileContent);
 
         $stmts = $this->nodeScopeAndMetadataDecorator->decorateNodesFromFile($file, $stmts);
         $file->hydrateStmtsAndTokens($stmts, $stmts, []);