--- a/src/Application/FileProcessor.php
+++ b/src/Application/FileProcessor.php
@@ -175,7 +175,7 @@ final class FileProcessor
     private function parseFileNodes(File $file): void
     {
         // store tokens by absolute path, so we don't have to print them right now
-        $stmtsAndTokens = $this->rectorParser->parseFileContentToStmtsAndTokens($file->getFileContent());
+        $stmtsAndTokens = $this->rectorParser->parseFileContentToStmtsAndTokens($file->getOriginalFileContent());
 
         $oldStmts = $stmtsAndTokens->getStmts();
         $oldTokens = $stmtsAndTokens->getTokens();