--- a/src/PhpParser/Parser/RectorParser.php
+++ b/src/PhpParser/Parser/RectorParser.php
@@ -18,6 +18,8 @@ final class RectorParser
     }
 
     /**
+     * @api used by rector-symfony
+     *
      * @return Stmt[]
      */
     public function parseFile(string $filePath): array