@staabm @TomasVotruba this should improve performance with avoid double read file.

This PR read file once, use on both parsing and filling File object by using `parseString()` method.