@samsonasik 2023-09-28
@staabm I tested run rector on rector-src, it seems faster:

**Before**

```
➜  rector-src git:(main) time bin/rector process --clear-cache
 2086/2086 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        

bin/rector process --clear-cache  219.56s user 10.26s system 713% cpu 32.191 total
```

**After**

```
➜  rector-src git:(dx-only-content) time bin/rector process --clear-cache
 2086/2086 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        

bin/rector process --clear-cache  216.03s user 8.79s system 715% cpu 31.410 total
```

It probably will more visible if the files has large content.