@TomasVotruba 2023-09-28
Not sure if 2 secs are worth it, but lets give it a go to cleanup code 👍